Refactor to pass parsl_config.py and workflow_config.py as a ConfigMap separate from the container image. Also run the script automatically as a Kubernetes job instead of creating a deployment where we have to start the script manually.
